ai-call-agent-resilience/requirements.txt
requests==2.31.0
gspread==5.12.0
oauth2client==4.1.3
python-telegram-bot==20.7
python-dotenv==1.0.0
